remove Context dependency from annotations
