+++ legate/core/runtime.py
@@ -15,6 +15,7 @@
 from __future__ import annotations
 
 import gc
+import inspect
 import math
 import struct
 import sys
@@ -22,7 +23,17 @@ import weakref
 from collections import deque
 from dataclasses import dataclass
 from types import ModuleType
-from typing import TYPE_CHECKING, Any, Deque, List, Optional, TypeVar, Union
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Callable,
+    Deque,
+    List,
+    Optional,
+    Protocol,
+    TypeVar,
+    Union,
+)
 
 from legion_top import add_cleanup_item, top_level
 
@@ -78,6 +89,47 @@ _sizeof_size_t = ffi.sizeof("size_t")
 assert _sizeof_size_t == 4 or _sizeof_size_t == 8
 
 _LEGATE_FIELD_ID_BASE = 1000
+
+
+class AnyCallable(Protocol):
+    def __call__(self, *args: Any, **kwargs: Any) -> Any:
+        ...
+
+
+def caller_frameinfo() -> str:
+    frame = inspect.currentframe()
+    if frame is None or frame.f_back is None or frame.f_back.f_back is None:
+        return "<unknown>"
+    frame = frame.f_back.f_back
+    return f"{frame.f_code.co_filename}:{frame.f_lineno}"
+
+
+class LibraryAnnotations:
+    def __init__(self) -> None:
+        self._entries: dict[str, str] = {}
+        self._provenance: Union[str, None] = None
+
+    @property
+    def provenance(self) -> Optional[str]:
+        return self._provenance
+
+    def set_provenance(self, provenance: str) -> None:
+        self._provenance = provenance
+
+    def reset_provenance(self) -> None:
+        self._provenance = None
+
+    def update(self, **kwargs: Any) -> None:
+        self._entries.update(**kwargs)
+
+    def remove(self, key: str) -> None:
+        del self._entries[key]
+
+    def __repr__(self) -> str:
+        pairs = [f"{key},{value}" for key, value in self._entries.items()]
+        if self._provenance is not None:
+            pairs.append(f"Provenance,{self._provenance}")
+        return "|".join(pairs)
 
 
 # A helper class for doing field management with control replication
@@ -1073,6 +1125,7 @@ class Runtime:
         )
 
         self._pending_exceptions: list[PendingException] = []
+        self._annotations: list[LibraryAnnotations] = [LibraryAnnotations()]
 
     @property
     def legion_runtime(self) -> legion.legion_runtime_t:
@@ -1183,6 +1236,35 @@ class Runtime:
     @property
     def field_match_manager(self) -> FieldMatchManager:
         return self._field_match_manager
+
+    @property
+    def annotation(self) -> LibraryAnnotations:
+        """
+        Returns the current set of annotations. Provenance string is one
+        entry in the set.
+
+        Returns
+        -------
+        LibraryAnnotations
+            Library annotations
+        """
+        return self._annotations[-1]
+
+    def get_all_annotations(self) -> str:
+        return str(self.annotation)
+
+    @property
+    def provenance(self) -> Optional[str]:
+        """
+        Returns the current provenance string. Attached to every operation
+        issued with the context.
+
+        Returns
+        -------
+        str or None
+            Provenance string
+        """
+        return self.annotation.provenance
 
     def register_library(self, library: Library) -> Context:
         """
@@ -1736,6 +1818,86 @@ class Runtime:
         for pending in pending_exceptions:
             pending.raise_exception()
 
+    def set_provenance(self, provenance: str) -> None:
+        """
+        Sets a new provenance string
+
+        Parameters
+        ----------
+        provenance : str
+            Provenance string
+        """
+        self._annotations[-1].set_provenance(provenance)
+
+    def reset_provenance(self) -> None:
+        """
+        Clears the provenance string that is currently set
+        """
+        self._annotations[-1].reset_provenance()
+
+    def push_provenance(self, provenance: str) -> None:
+        """
+        Pushes a provenance string to the stack
+
+        Parameters
+        ----------
+        provenance : str
+            Provenance string
+        """
+        self._annotations.append(LibraryAnnotations())
+        self.set_provenance(provenance)
+
+    def pop_provenance(self) -> None:
+        """
+        Pops the provenance string on top the stack
+        """
+        if len(self._annotations) == 1:
+            raise ValueError("Provenance stack underflow")
+        self._annotations.pop(-1)
+
+    def track_provenance(
+        self, func: AnyCallable, nested: bool = False
+    ) -> AnyCallable:
+        """
+        Wraps a function with provenance tracking. Provenance of each operation
+        issued within the wrapped function will be tracked automatically.
+
+        Parameters
+        ----------
+        func : AnyCallable
+            Function to wrap
+
+        nested : bool
+            If ``True``, each invocation to a wrapped function within another
+            wrapped function updates the provenance string. Otherwise, the
+            provenance is tracked only for the outermost wrapped function.
+
+        Returns
+        -------
+        AnyCallable
+            Wrapped function
+        """
+        if nested:
+
+            def wrapper(*args: Any, **kwargs: Any) -> Any:
+                self.push_provenance(caller_frameinfo())
+                result = func(*args, **kwargs)
+                self.pop_provenance()
+                return result
+
+        else:
+
+            def wrapper(*args: Any, **kwargs: Any) -> Any:
+                if self.provenance is None:
+                    self.set_provenance(caller_frameinfo())
+                    result = func(*args, **kwargs)
+                    self.reset_provenance()
+                else:
+                    result = func(*args, **kwargs)
+                return result
+
+        return wrapper
+
 
 runtime: Runtime = Runtime(core_library)
 
@@ -1800,3 +1962,55 @@ def get_legate_runtime() -> Runtime:
         Legate runtime object
     """
     return runtime
+
+
+def track_provenance(
+    nested: bool = False,
+) -> Callable[[AnyCallable], AnyCallable]:
+    """
+    Decorator that adds provenance tracking to functions. Provenance of each
+    operation issued within the wrapped function will be tracked automatically.
+
+    Parameters
+    ----------
+    nested : bool
+        If ``True``, each invocation to a wrapped function within another
+        wrapped function updates the provenance string. Otherwise, the
+        provenance is tracked only for the outermost wrapped function.
+
+    Returns
+    -------
+    Decorator
+        Function that takes a function and returns a one with provenance
+        tracking
+
+    See Also
+    --------
+    legate.core.runtime.Runtime.track_provenance
+    """
+
+    def decorator(func: AnyCallable) -> AnyCallable:
+        return runtime.track_provenance(func, nested=nested)
+
+    return decorator
+
+
+class Annotation:
+    def __init__(self, pairs: dict[str, str]) -> None:
+        """
+        Constructs a new annotation object
+
+        Parameters
+        ----------
+        pairs : dict[str, str]
+            Annotations as key-value pairs
+        """
+        self._annotation = runtime.annotation
+        self._pairs = pairs
+
+    def __enter__(self) -> None:
+        self._annotation.update(**self._pairs)
+
+    def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
+        for key in self._pairs.keys():
+            self._annotation.remove(key)
